refactor: distribute test_bugfixes.py into topic-specific test files
Each regression test from the June 2026 bug-fix batch now lives in its
natural home:
- Score judge response_schema tests → test_judge_response_schema.py
- Server path traversal / secret tests → test_file_uri.py
- CrossJudge credentials + compare_judges n_compared → test_cross_judge.py
- Duplicate scenario name detection → test_scenario_data.py
- Version single-sourcing → test_basic.py
- strip_thinking dangling tag → test_strip_thinking.py
- run_async ERROR isolation → test_model_auditor.py
- Experiment resume ERROR retry → test_resumable_experiments.py

All 569 tests still pass.
