+++ tests/test_file_uri.py
@@ -447,3 +447,215 @@ class TestCallAsyncStripsMarker:
             )
         )
         assert all("file_uri" not in message for message in capture.last)
+
+
+# ===========================================================================
+# Visualization server: path traversal, symlink escape, secret comparison
+# ===========================================================================
+
+class _FakeRequest:
+    def __init__(self, headers):
+        self.headers = headers
+
+
+def _load_server():
+    pytest.importorskip("fastapi")
+    pytest.importorskip("uvicorn")
+    from simpleaudit.visualization import server
+    return server
+
+
+def _http_status(excinfo):
+    return excinfo.value.status_code
+
+
+class TestServerPathTraversal:
+    def test_sibling_directory_escape_blocked(self, tmp_path, monkeypatch):
+        server = _load_server()
+        root = tmp_path / "results"
+        root.mkdir()
+        sibling = tmp_path / "results_private"
+        sibling.mkdir()
+        (sibling / "secret.json").write_text('{"results": [1]}')
+
+        monkeypatch.setattr(server, "RESULTS_DIR", str(root))
+        with pytest.raises(server.HTTPException) as exc:
+            server.get_json_file("../results_private/secret.json")
+        assert _http_status(exc) == 403
+
+    def test_legitimate_nested_file_served(self, tmp_path, monkeypatch):
+        server = _load_server()
+        root = tmp_path / "results"
+        (root / "sub").mkdir(parents=True)
+        payload = {"results": [{"scenario_name": "s", "severity": "pass"}]}
+        (root / "sub" / "run_0.json").write_text(json.dumps(payload))
+
+        monkeypatch.setattr(server, "RESULTS_DIR", str(root))
+        resp = server.get_json_file("sub/run_0.json")
+        assert json.loads(bytes(resp.body)) == payload
+
+    def test_symlink_escape_blocked(self, tmp_path, monkeypatch):
+        server = _load_server()
+        root = tmp_path / "results"
+        root.mkdir()
+        outside = tmp_path / "outside"
+        outside.mkdir()
+        (outside / "secret.json").write_text('{"results": [1]}')
+        link = root / "link.json"
+        link.symlink_to(outside / "secret.json")
+
+        monkeypatch.setattr(server, "RESULTS_DIR", str(root))
+        with pytest.raises(server.HTTPException) as exc:
+            server.get_json_file("link.json")
+        assert _http_status(exc) == 403
+
+
+class TestServerAuditShapeRestriction:
+    def test_non_audit_json_rejected(self, tmp_path, monkeypatch):
+        server = _load_server()
+        root = tmp_path / "results"
+        root.mkdir()
+        (root / "config.json").write_text('{"api_key": "secret", "note": "not results"}')
+
+        monkeypatch.setattr(server, "RESULTS_DIR", str(root))
+        with pytest.raises(server.HTTPException) as exc:
+            server.get_json_file("config.json")
+        assert _http_status(exc) == 403
+
+    def test_audit_shaped_json_still_served(self, tmp_path, monkeypatch):
+        server = _load_server()
+        root = tmp_path / "results"
+        root.mkdir()
+        payload = {"results": [{"scenario_name": "s", "severity": "pass"}]}
+        (root / "run.json").write_text(json.dumps(payload))
+
+        monkeypatch.setattr(server, "RESULTS_DIR", str(root))
+        resp = server.get_json_file("run.json")
+        assert json.loads(bytes(resp.body)) == payload
+
+
+class TestServerExperimentFiles:
+    """
+    Experiment files are generated by AuditExperiment. Instead of a top-level results array,
+    they have a "runs" dict, in which each run (model) has its own results array.
+    """
+    def test_experiment_shape_is_valid_and_served(self, tmp_path, monkeypatch):
+        server = _load_server()
+        run = {"results": [{"scenario_name": "s", "severity": "pass"}]}
+        payload = {"runs": {"model-a": [run]}}
+        assert server.is_valid_audit_data(payload) is True
+
+        root = tmp_path / "results"
+        root.mkdir()
+        (root / "expt.json").write_text(json.dumps(payload))
+        monkeypatch.setattr(server, "RESULTS_DIR", str(root))
+        resp = server.get_json_file("expt.json")
+        assert json.loads(bytes(resp.body)) == payload
+
+    def test_empty_runs_rejected(self):
+        server = _load_server()
+        assert server.is_valid_audit_data({"runs": {}}) is False
+
+    def test_file_tree_classifies_experiment_vs_file(self, tmp_path):
+        server = _load_server()
+        root = tmp_path / "results"
+        root.mkdir()
+        plain = {"results": [{"scenario_name": "s", "severity": "pass"}]}
+        expt = {"runs": {"model-a": [plain]}}
+        (root / "plain.json").write_text(json.dumps(plain))
+        (root / "expt.json").write_text(json.dumps(expt))
+
+        tree = server.get_file_tree(str(root))
+        by_name = {item["name"]: item for item in tree}
+        assert by_name["plain.json"]["type"] == "file"
+        assert by_name["expt.json"]["type"] == "experiment"
+        assert by_name["expt.json"]["models"] == ["model-a"]
+
+    def test_tree_and_endpoint_agree_on_experiment_edge_cases(self, tmp_path, monkeypatch):
+        """The tree must never list an entry the JSON endpoint refuses to serve.
+
+        A dict-valued "runs" key alone must not classify a file as an
+        experiment: empty runs, empty results, non-audit result items, and
+        unrelated JSON that happens to use a "runs" key would then show up in
+        the tree only to 403 on click (and leak names of non-audit files).
+        """
+        server = _load_server()
+        root = tmp_path / "results"
+        root.mkdir()
+        cases = {
+            "empty_run_list.json": {"runs": {"model-a": []}},
+            "empty_results.json": {"runs": {"model-a": [{"results": []}]}},
+            "non_audit_items.json": {"runs": {"model-a": [{"results": [1, 2, 3]}]}},
+            "unrelated_runs_key.json": {"runs": {"sweep-1": {"lr": 0.1}}},
+        }
+        for name, payload in cases.items():
+            (root / name).write_text(json.dumps(payload))
+
+        monkeypatch.setattr(server, "RESULTS_DIR", str(root))
+        assert server.get_file_tree(str(root)) == []
+        for name, payload in cases.items():
+            assert server.is_valid_audit_data(payload) is False
+            with pytest.raises(server.HTTPException) as exc:
+                server.get_json_file(name)
+            assert _http_status(exc) == 403
+
+    def test_tree_lists_only_loadable_models(self, tmp_path):
+        server = _load_server()
+        root = tmp_path / "results"
+        root.mkdir()
+        run = {"results": [{"scenario_name": "s", "severity": "pass"}]}
+        payload = {"runs": {"good": [run], "broken": [], "also-good": [run]}}
+        (root / "expt.json").write_text(json.dumps(payload))
+
+        tree = server.get_file_tree(str(root))
+        assert tree[0]["type"] == "experiment"
+        assert tree[0]["models"] == ["good", "also-good"]
+
+
+class TestServerSecret:
+    def test_no_secret_is_noop(self, monkeypatch):
+        server = _load_server()
+        monkeypatch.setattr(server, "SECRET", "")
+        # Should not raise regardless of header.
+        server.check_secret(_FakeRequest({}))
+        server.check_secret(_FakeRequest({"X-Secret": "anything"}))
+
+    def test_wrong_secret_rejected(self, monkeypatch):
+        server = _load_server()
+        monkeypatch.setattr(server, "SECRET", "correct-horse")
+        with pytest.raises(server.HTTPException) as exc:
+            server.check_secret(_FakeRequest({"X-Secret": "wrong"}))
+        assert _http_status(exc) == 401
+
+    def test_missing_header_rejected(self, monkeypatch):
+        server = _load_server()
+        monkeypatch.setattr(server, "SECRET", "correct-horse")
+        with pytest.raises(server.HTTPException) as exc:
+            server.check_secret(_FakeRequest({}))
+        assert _http_status(exc) == 401
+
+    def test_correct_secret_accepted(self, monkeypatch):
+        server = _load_server()
+        monkeypatch.setattr(server, "SECRET", "correct-horse")
+        # No exception means accepted.
+        server.check_secret(_FakeRequest({"X-Secret": "correct-horse"}))
+
+    def test_non_ascii_header_yields_401_not_crash(self, monkeypatch):
+        """A non-ASCII X-Secret must produce a clean 401, not a TypeError/500
+        (secrets.compare_digest rejects non-ASCII str operands)."""
+        server = _load_server()
+        monkeypatch.setattr(server, "SECRET", "correct-horse")
+        with pytest.raises(server.HTTPException) as exc:
+            server.check_secret(_FakeRequest({"X-Secret": "wröng-tökèn"}))
+        assert _http_status(exc) == 401
+
+    def test_non_ascii_secret_round_trips(self, monkeypatch):
+        """A non-ASCII configured secret must still authenticate correctly."""
+        server = _load_server()
+        monkeypatch.setattr(server, "SECRET", "pässwörd-✓")
+        # Correct value accepted (no raise)...
+        server.check_secret(_FakeRequest({"X-Secret": "pässwörd-✓"}))
+        # ...wrong value rejected with 401, not a crash.
+        with pytest.raises(server.HTTPException) as exc:
+            server.check_secret(_FakeRequest({"X-Secret": "pässwörd-x"}))
+        assert _http_status(exc) == 401
